general database functions and testing site
