+++ web_app/test.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <script src="https://www.gstatic.com/firebasejs/4.8.1/firebase.js"></script>
+    <script src="https://cdn.firebase.com/libs/firebaseui/2.5.1/firebaseui.js"></script>
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+    <script src="js/firebase_auth.js"></script>
+    <script src="js/main.js"></script>
+    <link type="text/css" rel="stylesheet" href="https://cdn.firebase.com/libs/firebaseui/2.5.1/firebaseui.css" />
+    <link type="text/css" rel="stylesheet" href="styles.css">
+    <title>Test</title>
+</head>
+<body>
+    <p id ="color">1<p>
+    <p id ="description">asdf<p>
+    <p id ="name">asdf<p>
+    <p id ="priority">1<p>
+    <p id ="time">get current time<p>
+    <p id ="user">whoisthis<p>
+    <div id="firebaseui-auth-container"></div>
+</body>
+</html>
